refactor: ItemSearchDto를 web 계층에서 domain 계층으로 이동
