+++ src/main/java/shop/tryit/domain/item/ItemRepository.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import shop.tryit.web.item.ItemSearchDto;
 
 public interface ItemRepository {
