+++ src/main/java/shop/tryit/domain/item/ItemService.java
@@ -6,7 +6,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shop.tryit.web.item.ItemSearchDto;
 
 @Service
 @RequiredArgsConstructor
