+++ src/test/java/shop/tryit/domain/item/ItemServiceTests.java
@@ -14,7 +14,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
-import shop.tryit.web.item.ItemSearchDto;
 
 @Transactional
 @SpringBootTest
